requirements.txt
tensorflow==2.7.0
tensorflow-hub==0.12.0
matplotlib==3.5.1
Pillow==9.0.0
numpy==1.21.5
